+++ js/defaults.js
@@ -18,6 +18,10 @@ var defaults = {
 
 	modules: [
 		{
+			module: "updatenotification",
+			position: "top_center"
+		},
+		{
 			module: "helloworld",
 			position: "upper_third",
 			classes: "large thin",